change SM configuration syntax to no longer use WithConfiguration(..) but use In(..) / DefineHierarchyOn(..) directly on the StateMachineDefinitionBuilder
This is the same syntax as before version 5.0. The only difference is it is no longer directly on a ASM / PSM .. but on a StateMachineDefinitionBuilder.

The reason for abandoning the WithConfiguration(..) syntax is that it increases nesting. Which for bigger setups reduces readability.

A semantical change to the WithConfiguration(..) syntax is that configuration state is mutated immediately with each In(..) / DefineHierarchyOn(..) instead of only when Build() is executed.