change StateLogic to no longer hold StateDefinitionsDictionary as instance var but receive it via method parameter
Ultimatively to goal is to get towards a more functional style. Differentiating between classes that have contain only logic and classes that contain only data.
StateDefinitionDictionary is clearly a data class and should therefore not be held as instance var.
As another example TransitionLogic is clearly a logic class and stateless (considering only data) and can therefore be held as an instance var.